requirements.txt
kafka-python
faker
Flask